==> 01. Log Extract and Parsing.ipynb <==
## Import Library

import tarfile
import os
import gzip
import shutil
import numpy as np
import pandas as pd
import json
import re
from collections import defaultdict
from tqdm import tqdm
from logparser import Spell, Drain, Drain_bgl
from timeit import default_timer as timer

## Extract Log File

#Dataset Tar File
hdfs_file = "datasets/HDFS_1.tar.gz"
bgl_file = "datasets/BGL.tar.gz"
tbird_file = "datasets/tbird2.gz"

#Make Output Directory
_path = {"hdfs":"datasets/hdfs/", "tbird":"datasets/tbird/", "bgl":"datasets/bgl/"}

for key, path in _path.items() :
  try:
      os.mkdir(path)
  except OSError as error:
      pass

#Extract HDFS LogFile
start = timer()
if os.path.exists("datasets/hdfs/HDFS.log") == False:
    print("Extracting...")
    with tarfile.open(hdfs_file, 'r') as tar_ref:
        tar_ref.extractall(_path["hdfs"])
else : print("Already Extracted")
end = timer()
print("time elapsed : {:.2f}s".format(end-start)) 

#Extract bgl LogFile
start = timer()
if os.path.exists("datasets/hdfs/HDFS.log") == False:
    print("Extracting...")
    with tarfile.open(bgl_file, 'r') as tar_ref:
        tar_ref.extractall(_path["bgl"])
else : print("Already Extracted")
end = timer()
print("time elapsed : {:.2f}s".format(end-start)) 

#Extract tbird LogFile
start = timer()
if os.path.exists("datasets/hdfs/HDFS.log") == False:
    print("Extracting...")
    with gzip.open(tbird_file, 'rb') as f_in:
        with open(_path["tbird"] + 'tbird.log', 'wb') as f_out:
            shutil.copyfileobj(f_in, f_out)
else : print("Already Extracted")
end = timer()
print("time elapsed : {:.2f}s".format(end-start)) 

## Parse Log File

#### Helper Function

def sliding_window(raw_data, para):
    """
    split logs into sliding windows/session
    :param raw_data: dataframe columns=[timestamp, label, eventid, time duration]
    :param para:{window_size: seconds, step_size: seconds}
    :return: dataframe columns=[eventids, time durations, label]
    """
    log_size = raw_data.shape[0]
    label_data, time_data = raw_data.iloc[:, 1], raw_data.iloc[:, 0]
    logkey_data, deltaT_data = raw_data.iloc[:, 2], raw_data.iloc[:, 3]
    new_data = []
    start_end_index_pair = set()

    start_time = time_data[0]
    end_time = start_time + para["window_size"]
    start_index = 0
    end_index = 0

    # get the first start, end index, end time
    for cur_time in time_data:
        if cur_time < end_time:
            end_index += 1
        else:
            break

    start_end_index_pair.add(tuple([start_index, end_index]))

    # move the start and end index until next sliding window
    num_session = 1
    while end_index < log_size:
        start_time = start_time + para['step_size']
        end_time = start_time + para["window_size"]
        for i in range(start_index, log_size):
            if time_data[i] < start_time:
                i += 1
            else:
                break
        for j in range(end_index, log_size):
            if time_data[j] < end_time:
                j += 1
            else:
                break
        start_index = i
        end_index = j

        # when start_index == end_index, there is no value in the window
        if start_index != end_index:
            start_end_index_pair.add(tuple([start_index, end_index]))

        num_session += 1
        if num_session % 1000 == 0:
            print("process {} time window".format(num_session), end='\r')

    for (start_index, end_index) in start_end_index_pair:
        dt = deltaT_data[start_index: end_index].values
        dt[0] = 0
        new_data.append([
            time_data[start_index: end_index].values,
            max(label_data[start_index:end_index]),
            logkey_data[start_index: end_index].values,
            dt
        ])

    assert len(start_end_index_pair) == len(new_data)
    print('there are %d instances (sliding windows) in this dataset\n' % len(start_end_index_pair))
    return pd.DataFrame(new_data, columns=raw_data.columns)

### - HDFS

#PATH
log_format = '<Date> <Time> <Pid> <Level> <Component>: <Content>' #HDFS Log Format
input_dir  = _path["hdfs"]
output_dir = _path["hdfs"] + 'output/'
log_file   = "HDFS.log"
log_structured_file = output_dir + log_file + "_structured.csv"
log_templates_file = output_dir + log_file + "_templates.csv"
log_sequence_file = output_dir + "hdfs_sequence.csv"
log_content_file = output_dir + "hdfs_content.csv"
log_results_file = output_dir + "hdfs_labeled.csv"
log_results_content_file = output_dir + "hdfs_content_labeled.csv"
logkey_content_file = output_dir + "hdfs_content_logkey.csv"

def parser(input_dir, output_dir, log_file, log_format):
  regex = [
      r"(?<=blk_)[-\d]+", #Regex for block_id
      r'\d+\.\d+\.\d+\.\d+',  #Regex for  IP
      r"(/[-\w]+)+",  #Regex for file path
      #r'(?<=[^A-Za-z0-9])(\-?\+?\d+)(?=[^A-Za-z0-9])|[0-9]+$',  # Numbers
  ]
  # hyper parameter di set berdasarkan  http://jmzhu.logpai.com/pub/pjhe_icws2017.pdf
  st = 0.5  # Similarity threshold
  depth = 5  # Depth of all leaf nodes

  parser = Drain.LogParser(log_format, indir=input_dir, outdir=output_dir, depth=depth, st=st, rex=regex, keep_para=False)
  parser.parse(log_file)

parser(input_dir, output_dir, log_file, log_format)

def mapping():
    log_temp = pd.read_csv(log_templates_file)
    log_temp.sort_values(by = ["Occurrences"], ascending=False, inplace=True)
    log_temp_dict = {event: idx+1 for idx , event in enumerate(list(log_temp["EventId"])) }
    print(log_temp_dict)
    with open (output_dir + "hdfs_log_templates.json", "w") as f:
        json.dump(log_temp_dict, f)

mapping()

def hdfs_content_append(log_file, window='session'):
    assert window == 'session', "Only window=session is supported for HDFS dataset."
    print("Loading", log_file)
    df = pd.read_csv(log_file, engine='c',
            na_filter=False, memory_map=True, dtype={'Date':object, "Time": object})

    with open(output_dir + "hdfs_log_templates.json", "r") as f:
        event_num = json.load(f)
    df["EventId"] = df["EventId"].apply(lambda x: event_num.get(x, -1))

    data_dict = defaultdict(list)
    for idx, row in tqdm(df.iterrows()):
        blkId_list = re.findall(r'(blk_-?\d+)', row['Content'])
        blkId_set = set(blkId_list)
        for blk_Id in blkId_set:
            data_dict[blk_Id].append(row["EventTemplate"])

    data_df = pd.DataFrame(list(data_dict.items()), columns=['BlockId', 'ContentSequence'])
    data_df.to_csv(log_content_file, index=None)
    print("hdfs sampling done")

hdfs_content_append(log_structured_file)

def hdfs_sequencing(log_file, window='session'):
    assert window == 'session', "Only window=session is supported for HDFS dataset."
    print("Loading", log_file)
    df = pd.read_csv(log_file, engine='c',
            na_filter=False, memory_map=True, dtype={'Date':object, "Time": object})

    # with open(output_dir + "hdfs_log_templates.json", "r") as f:
    #     event_num = json.load(f)
    # df["EventId"] = df["EventId"].apply(lambda x: event_num.get(x, -1))

    data_dict = defaultdict(list)
    for idx, row in tqdm(df.iterrows()):
        blkId_list = re.findall(r'(blk_-?\d+)', row['Content'])
        blkId_set = set(blkId_list)
        for blk_Id in blkId_set:
            data_dict[blk_Id].append(row["EventId"])

    data_df = pd.DataFrame(list(data_dict.items()), columns=['BlockId', 'EventSequence'])
    data_df.to_csv(log_sequence_file, index=None)
    print("hdfs sampling done")

hdfs_sequencing(log_structured_file)

def generate_dataset(hdfs_sequence_file, results=log_results_file):
    blk_label_dict = {}
    blk_label_file = os.path.join(input_dir, "anomaly_label.csv")
    blk_df = pd.read_csv(blk_label_file)
    for _ , row in tqdm(blk_df.iterrows()):
        # blk_label_dict[row["BlockId"]] = row["Label"]
        blk_label_dict[row["BlockId"]] = 1 if row["Label"] == "Anomaly" else 0

    seq = pd.read_csv(hdfs_sequence_file)
    seq["Label"] = seq["BlockId"].apply(lambda x: blk_label_dict.get(x)) #Tambah label ke tiap sequence dari tiap blockid
    seq.to_csv(results, index=None)

generate_dataset(log_sequence_file)

generate_dataset(log_content_file, log_results_content_file)

df = pd.read_csv(log_results_file)
df.head()

df["Label"].value_counts()

df = pd.read_csv(log_results_content_file)
df.head()

df["Label"].value_counts()

df_logkey = pd.read_csv(log_results_file)

df_content = pd.read_csv(log_results_content_file)

merged_df = pd.merge(df, df_content[["BlockId", "ContentSequence"]], on="BlockId", how="inner")
merged_df = merged_df[["BlockId", "EventSequence", "ContentSequence", "Label"]]

merged_df.head()

merged_df.isnull().sum()

merged_df["Label"].value_counts()

merged_df.to_csv(logkey_content_file, index=None)

### - BGL

#PATH and VAR
data_dir = _path["bgl"]
output_dir = _path["bgl"] + "output/"
log_file = "BGL.log"
log_structured_file  = output_dir + log_file + "_structured.csv"
log_results_file = output_dir + "bgl_time_windowed.csv"
log_results_content_file = output_dir + "bgl_time_windowed_content.csv"
logkey_content_file = output_dir + "bgl_content_logkey.csv"
window_size = 5
step_size = 1

def parse_log(input_dir, output_dir, log_file, parser_type):
    log_format = '<Label> <Id> <Date> <Code1> <Time> <Code2> <Component1> <Component2> <Level> <Content>'
    regex = [
        r'(0x)[0-9a-fA-F]+', #hexadecimal
        r'\d+.\d+.\d+.\d+',
        # r'/\w+( )$'
        r'\d+'
    ]
    keep_para = False
    if parser_type == "drain":
        # the hyper parameter is set according to http://jmzhu.logpai.com/pub/pjhe_icws2017.pdf
        st = 0.3  # Similarity threshold
        depth = 3  # Depth of all leaf nodes
        parser = Drain_bgl.LogParser(log_format, indir=input_dir, outdir=output_dir, depth=depth, st=st, rex=regex, keep_para=keep_para)
        parser.parse(log_file)
    elif parser_type == "spell":
        tau = 0.55
        parser = Spell.LogParser(indir=data_dir, outdir=output_dir, log_format=log_format, tau=tau, rex=regex, keep_para=keep_para)
        parser.parse(log_file)

parse_log(data_dir, output_dir, log_file, 'drain')

df = pd.read_csv(log_structured_file)

df["Label"] = df["Label"].apply(lambda x: int(x != "-"))
df['datetime'] = pd.to_datetime(df['Time'], format='%Y-%m-%d-%H.%M.%S.%f')
df['timestamp'] = df["datetime"].values.astype(np.int64) // 10 ** 9
df['deltaT'] = df['datetime'].diff() / np.timedelta64(1, 's')
df['deltaT'] = df['deltaT'].fillna(0)

start = timer()
bgl_df = sliding_window(df[["timestamp", "Label", "EventId", "deltaT"]],
                                para={"window_size": int(window_size)*60, "step_size": int(step_size) * 60}
                                )
end = timer()
print("time elapsed : {:.2f}s".format(end-start)) 

bgl_df.to_csv(log_results_file, index = None)

start = timer()
bgl_df = sliding_window(df[["timestamp", "Label", "EventTemplate", "deltaT"]],
                                para={"window_size": int(window_size)*60, "step_size": int(step_size) * 60}
                                )
end = timer()
print("time elapsed : {:.2f}s".format(end-start)) 

bgl_df.to_csv(log_results_content_file, index = None)

df = pd.read_csv(log_results_file)
df.head()

df["Label"].value_counts()

df = pd.read_csv(log_results_content_file)
df.head()

df["Label"].value_counts()

df_logkey = pd.read_csv(log_results_file)
df_content = pd.read_csv(log_results_content_file)

merged_df = pd.merge(df_logkey, df_content[["timestamp", "EventTemplate"]], on="timestamp", how="inner")
merged_df = merged_df[["timestamp", "EventId", "EventTemplate", "Label"]]

merged_df.head()

merged_df.isnull().sum()

merged_df["Label"].value_counts()

merged_df.to_csv(logkey_content_file, index=None)
print("Saved at : " + logkey_content_file)

### - Thunderbird

#PATH and VAR
raw_log = _path["tbird"] + "tbird.log"
sample_log = _path["tbird"] + "tbird_5M.log"
sample_window_size = 5*10**6
sample_step_size = 10**3
data_dir = _path["tbird"]
output_dir = _path["tbird"] + "output/"
log_file = "tbird_5M.log"
window_size = 1
step_size = 0.5
log_results_file = output_dir + "tbird_time_windowed_5M.csv"
log_results_content_file = output_dir + "tbird_time_windowed_5M_content.csv"
logkey_content_file = output_dir + "tbird_content_logkey.csv"

def sample_raw_data(data_file, output_file, sample_window_size, sample_step_size):
    # sample 1M by sliding window, abnormal rate is over 2%
    sample_data = []
    labels = []
    idx = 0

    # spirit dataset can start from the 2Mth line, as there are many abnormal lines gathering in the first 2M
    with open(data_file, 'r', errors='ignore') as f:
        for line in f:
            labels.append(line.split()[0] != '-')
            sample_data.append(line)

            if len(labels) == sample_window_size:
                abnormal_rate = sum(np.array(labels)) / len(labels)
                print(f"{idx + 1} lines, abnormal rate {abnormal_rate}")
                break

            idx += 1
            if idx % sample_step_size == 0:
                print(f"Process {round(idx/sample_window_size * 100,4)} % raw data", end='\r')

    with open(output_file, "w") as f:
        f.writelines(sample_data)

    print("Sampling done")

start = timer()
sample_raw_data(raw_log, sample_log, sample_window_size, sample_step_size)
end = timer()
print("time elapsed : {:.2f}s".format(end-start))

def parse_log(input_dir, output_dir, log_file, parser_type):
    log_format = '<Label> <Id> <Date> <Admin> <Month> <Day> <Time> <AdminAddr> <Content>'
    regex = [
        r'(0x)[0-9a-fA-F]+',  # hexadecimal
        r'\d+\.\d+\.\d+\.\d+',
        r'(?<=Warning: we failed to resolve data source name )[\w\s]+',
        r'\d+'
    ]
    keep_para = False
    if parser_type == "drain":
        # the hyper parameter is set according to http://jmzhu.logpai.com/pub/pjhe_icws2017.pdf
        st = 0.3  # Similarity threshold
        depth = 3  # Depth of all leaf nodes

        # Drain is modified
        parser = Drain.LogParser(log_format,
                                 indir=input_dir,
                                 outdir=output_dir,
                                 depth=depth,
                                 st=st,
                                 rex=regex,
                                 keep_para=keep_para, maxChild=1000)
        parser.parse(log_file)

    elif parser_type == "spell":
        tau = 0.35
        parser = Spell.LogParser(indir=data_dir,
                                 outdir=output_dir,
                                 log_format=log_format,
                                 tau=tau,
                                 rex=regex,
                                 keep_para=keep_para)
        parser.parse(log_file)

parse_log(data_dir, output_dir, log_file, "drain")

df = pd.read_csv(output_dir + "tbird_5M.log_structured.csv")

df["Label"] = df["Label"].apply(lambda x: int(x != "-"))
df['datetime'] = pd.to_datetime(df["Date"] + " " + df['Time'], format='%Y.%m.%d %H:%M:%S')
df['timestamp'] = df["datetime"].values.astype(np.int64) // 10 ** 9
df['deltaT'] = df['datetime'].diff() / np.timedelta64(1, 's')
df['deltaT'] = df['deltaT'].fillna(0)

start = timer()
window_df = sliding_window(df[["timestamp", "Label", "EventId", "deltaT"]], 
                           para={"window_size": float(1)*60, "step_size": float(0.5) * 60})
end = timer()
print("time elapsed : {:.2f}s".format(end-start))

window_df.to_csv(log_results_file, index = None)

start = timer()
window_df = sliding_window(df[["timestamp", "Label", "EventTemplate", "deltaT"]], 
                           para={"window_size": float(1)*60, "step_size": float(0.5) * 60})
end = timer()
print("time elapsed : {:.2f}s".format(end-start))

window_df.to_csv(log_results_content_file, index = None)

df = pd.read_csv(log_results_file)
df.head()

df["Label"].value_counts()

df["Label"].value_counts()

df = pd.read_csv(log_results_content_file)
df.head()

df["Label"].value_counts()

df_logkey = pd.read_csv(log_results_file)
df_content = pd.read_csv(log_results_content_file)

merged_df = pd.merge(df_logkey, df_content[["timestamp", "EventTemplate"]], on="timestamp", how="inner")
merged_df = merged_df[["timestamp", "EventId", "EventTemplate", "Label"]]

merged_df.head()

merged_df.isnull().sum()

merged_df["Label"].value_counts()

merged_df.to_csv(logkey_content_file, index=None)
print("Saved at : " + logkey_content_file)

### SANDBOX

df = pd.read_csv(log_results_file)

df.head()

df_content = pd.read_csv(log_results_content_file)

df_content.head()

merged_df = pd.merge(df, df_content[["timestamp", "EventTemplate"]], on="timestamp", how="inner")
merged_df = merged_df[["timestamp", "EventId", "EventTemplate", "Label"]]

merged_df.head()

merged_df.isnull().sum()
