==> 02. Create Training Test and Eval Datasets.ipynb <==
## Import Library

import pandas as pd
import os

## Path

hdfs_output_dir = "datasets/hdfs/output/"
hdfs_file = "hdfs_labeled.csv"
hdfs_content_file = "hdfs_content_labeled.csv"
hdfs_logkey_content = "hdfs_content_logkey.csv"
hdfs_dataset = hdfs_output_dir + hdfs_file
hdfs_template = hdfs_output_dir + "HDFS.log_templates.csv"
hdfs_dataset_dir = hdfs_output_dir + "dataset/"

bgl_output_dir = "datasets/bgl/output/"
bgl_file = "bgl_time_windowed.csv"
bgl_content_file = "bgl_time_windowed_content.csv"
bgl_logkey_content = "bgl_content_logkey.csv"
bgl_dataset = bgl_output_dir + bgl_file
bgl_template = bgl_output_dir + "BGL.log_templates.csv"
bgl_dataset_dir = bgl_output_dir + "dataset/"

tbird_output_dir = "datasets/tbird/output/"
tbird_file = "tbird_time_windowed_5M.csv"
tbird_content_file = "tbird_time_windowed_5M_content.csv"
tbird_logkey_content = "tbird_content_logkey.csv"
tbird_dataset = tbird_output_dir + tbird_file
tbird_template = tbird_output_dir + "tbird_5M.log_templates.csv"
tbird_dataset_dir = tbird_output_dir + "dataset/"

output_path = [hdfs_output_dir, bgl_output_dir, tbird_output_dir]
for path in output_path :
  try:
      os.mkdir(path + "dataset/")
  except OSError as error:
      pass

## Generate Train, Test, and Eval Datasets

def generate_train_test(log_dir, log_file, output_dir, mode=None, content=False, n=5000, test_n=1000):
    df = pd.read_csv(log_dir + log_file)
    if mode == 'hdfs':
        if content : 
            df = df[['ContentSequence', 'Label']]
            df.rename(columns={'ContentSequence': 'text'}, inplace=True)
            df.rename(columns={'Label': 'label'}, inplace=True)
        else :
            df = df[['EventSequence', 'Label']]
            df.rename(columns={'EventSequence': 'text'}, inplace=True)
            df.rename(columns={'Label': 'label'}, inplace=True)
    else :
        if content : 
            df = df[['EventTemplate', 'Label']]
            df.rename(columns={'EventTemplate': 'text'}, inplace=True)
            df.rename(columns={'Label': 'label'}, inplace=True)
        else :
            df = df[['EventId', 'Label']]
            df.rename(columns={'EventId': 'text'}, inplace=True)
            df.rename(columns={'Label': 'label'}, inplace=True)
    print("===== Original Dataset =====")
    print(df['label'].value_counts())
        
    # train df  
    normal_seq = df[df["label"] == 0].sample(n, random_state=20)
    anomaly_seq = df[df["label"] == 1].sample(test_n, random_state=20)
    train_df = pd.concat([normal_seq, anomaly_seq], ignore_index=False, sort=False)
    print("\n===== Training Dataset =====")
    print(train_df['label'].value_counts())
    train_df.to_csv(output_dir + 'train_' + log_file, index = None)
    print('saved to : ' + output_dir + 'train_' + log_file)

    # eval df 
    df_eval = df.drop(train_df.index)
    normal_seq = df_eval[df_eval["label"] == 0].sample(test_n, random_state=20)
    anomaly_seq = df_eval[df_eval["label"] == 1].sample(test_n, random_state=20)
    eval_df = pd.concat([normal_seq, anomaly_seq], ignore_index=False, sort=False)
    print("\n===== Evaluate Dataset =====")
    print(eval_df['label'].value_counts())
    eval_df.to_csv(output_dir + 'eval_' + log_file, index = None)
    print('saved to : ' + output_dir + 'eval_' + log_file)

    # test df
    df_test = df_eval.drop(eval_df.index)
    normal_seq = df_test[df_test["label"] == 0].sample(test_n, random_state=20)
    anomaly_seq = df_test[df_test["label"] == 1].sample(test_n, random_state=20)
    test_df = pd.concat([normal_seq, anomaly_seq], ignore_index=False, sort=False)
    print("\n===== Testing Dataset =====")
    print(test_df['label'].value_counts())
    test_df.to_csv(output_dir + 'test_' + log_file, index = None)
    print('saved to : ' + output_dir + 'test_' + log_file)
    
    

### HDFS

generate_train_test(hdfs_output_dir, hdfs_file, hdfs_dataset_dir, mode='hdfs')

generate_train_test(hdfs_output_dir, hdfs_content_file, hdfs_dataset_dir, mode='hdfs', content=True)

### BGL

generate_train_test(bgl_output_dir, bgl_file, bgl_dataset_dir)

generate_train_test(bgl_output_dir, bgl_content_file, bgl_dataset_dir, content=True)

### Thunderbird

generate_train_test(tbird_output_dir, tbird_file, tbird_dataset_dir)

generate_train_test(tbird_output_dir, tbird_content_file, tbird_dataset_dir,content=True)

## Generate Logkey & Content : Train, Test, Eval Dataset

def generate_logkey_content(log_dir, log_file, output_dir, mode=None, n=5000, test_n=1000):
    df = pd.read_csv(log_dir + log_file)
    if mode == 'hdfs':
        df = df[['EventSequence', 'ContentSequence', 'Label']]
        df.rename(columns={'EventSequence': 'logkey'}, inplace=True)
        df.rename(columns={'ContentSequence': 'content'}, inplace=True)
        df.rename(columns={'Label': 'label'}, inplace=True)
    else :
        df = df[['EventId', 'EventTemplate', 'Label']]
        df.rename(columns={'EventId': 'logkey'}, inplace=True)
        df.rename(columns={'EventTemplate': 'content'}, inplace=True)
        df.rename(columns={'Label': 'label'}, inplace=True)
    
    print("===== Original Dataset =====")
    print(df['label'].value_counts())
        
    # train df  
    normal_seq = df[df["label"] == 0].sample(n, random_state=20)
    # anomaly_seq = df[df["label"] == 1].sample(test_n, random_state=20)
    # train_df = pd.concat([normal_seq, anomaly_seq], ignore_index=False, sort=False)
    train_df = normal_seq
    print("\n===== Training Dataset =====")
    print(train_df['label'].value_counts())
    train_df.to_csv(output_dir + 'lc_train_' + log_file, index = None)
    print('saved to : ' + output_dir + 'lc_train_' + log_file)

    # eval df 
    df_eval = df.drop(train_df.index)
    normal_seq = df_eval[df_eval["label"] == 0].sample(test_n, random_state=20)
    anomaly_seq = df_eval[df_eval["label"] == 1].sample(test_n, random_state=20)
    eval_df = pd.concat([normal_seq, anomaly_seq], ignore_index=False, sort=False)
    print("\n===== Evaluate Dataset =====")
    print(eval_df['label'].value_counts())
    eval_df.to_csv(output_dir + 'lc_eval_' + log_file, index = None)
    print('saved to : ' + output_dir + 'lc_eval_' + log_file)

    # test df
    df_test = df_eval.drop(eval_df.index)
    normal_seq = df_test[df_test["label"] == 0].sample(test_n, random_state=20)
    anomaly_seq = df_test[df_test["label"] == 1].sample(test_n, random_state=20)
    test_df = pd.concat([normal_seq, anomaly_seq], ignore_index=False, sort=False)
    print("\n===== Testing Dataset =====")
    print(test_df['label'].value_counts())
    test_df.to_csv(output_dir + 'lc_test_' + log_file, index = None)
    print('saved to : ' + output_dir + 'lc_test_' + log_file)

generate_logkey_content(hdfs_output_dir, hdfs_logkey_content, hdfs_dataset_dir, mode='hdfs')

generate_logkey_content(bgl_output_dir, bgl_logkey_content, bgl_dataset_dir)

generate_logkey_content(tbird_output_dir, tbird_logkey_content, tbird_dataset_dir)

## SANDBOX

### SANDBOX 1 Generate Train Test Dataset

df = pd.read_csv(hdfs_dataset)
df["Label"].value_counts()

df = df.drop("BlockId", axis='columns')
df.head()

df.rename(columns={'EventSequence': 'text'}, inplace=True)
df.rename(columns={'Label': 'label'}, inplace=True)
df.head()

#### Training Dataset

normal_seq = df[df["label"] == 0].sample(n=5000, random_state=1)
normal_seq

anomaly_seq = df[df["label"] == 1].sample(n=1000, random_state=1)
anomaly_seq

hdfs_df = pd.concat([normal_seq, anomaly_seq], ignore_index=False, sort=False)
hdfs_df

hdfs_df["label"].value_counts()

hdfs_df.to_csv(hdfs_output_dir + "dataset/hdfs_dataset_training.csv", index = None)

#### Evaluate Dataset

df_eval = df.drop(hdfs_df.index)
df_eval["label"].value_counts()

normal_eval = df_eval[df_eval["label"] == 0].sample(n=1000, random_state=1)
normal_eval

anomaly_eval = df_eval[df_eval["label"] == 1].sample(n=1000, random_state=1)
anomaly_eval

hdfs_eval = pd.concat([normal_eval, anomaly_eval], ignore_index=False, sort=False)
hdfs_eval

hdfs_eval["label"].value_counts()

hdfs_eval.to_csv(hdfs_output_dir + "dataset/hdfs_dataset_evaluate.csv", index = None)

#### Test Dataset

df_test = df_eval.drop(hdfs_eval.index)
df_test["label"].value_counts()

normal_test = df_test[df_test["label"] == 0].sample(n=1000, random_state=1)
normal_test

anomaly_test = df_test[df_test["label"] == 1].sample(n=1000, random_state=1)
anomaly_test

hdfs_test = pd.concat([normal_test, anomaly_test], ignore_index=False, sort=False)
hdfs_test

hdfs_test["label"].value_counts()

hdfs_test.to_csv(hdfs_output_dir + "dataset/hdfs_dataset_test.csv", index = None)

### SANDBOX 2 Generate Content Logkey
