==> 03. Create Tokenizer.ipynb <==
## Import Library

import pandas as pd
from datasets import Dataset
from transformers import AutoTokenizer
from timeit import default_timer as timer

hdfs_data_dir = "datasets/hdfs/output/"
hdfs_file = "hdfs_labeled.csv"
hdfs_content_file = "hdfs_content_labeled.csv"
hdfs_logkey_content = "hdfs_content_logkey.csv"
hdfs_output_dir = "tokenizer/hdfs"

bgl_data_dir = "datasets/bgl/output/"
bgl_file = "bgl_time_windowed.csv"
bgl_content_file = "bgl_time_windowed_content.csv"
bgl_logkey_content = "bgl_content_logkey.csv"
bgl_output_dir = "tokenizer/bgl"

tbird_data_dir = "datasets/tbird/output/"
tbird_file = "tbird_time_windowed_5M.csv"
tbird_content_file = "tbird_time_windowed_5M_content.csv"
tbird_logkey_content = "tbird_content_logkey.csv"
tbird_output_dir = "tokenizer/tbird"

## Generate Tokenizer

def generate_tokenizer(log_dir, log_file, output_dir, token, mode=None, content=False, vocab_size = 50265):
    start = timer()
    df = pd.read_csv(log_dir + log_file)
    if mode == 'hdfs':
        if content:
            df = df[["ContentSequence"]]
            df.rename(columns={'ContentSequence': 'text'}, inplace=True)
        else :
            df = df[["EventSequence"]]
            df.rename(columns={'EventSequence': 'text'}, inplace=True)
    else :
        if content:
            df = df[["EventTemplate"]]
            df.rename(columns={'EventTemplate': 'text'}, inplace=True)
        else :
            df = df[["EventId"]]
            df.rename(columns={'EventId': 'text'}, inplace=True)
    
    dataset = Dataset.from_pandas(df)
    batch_size = 1000
    
    def batch_iterator():
        for i in range(0, len(dataset), batch_size):
            yield dataset[i : i + batch_size]["text"]

    tokenizer = token.train_new_from_iterator(batch_iterator(), vocab_size)
    tokenizer.save_pretrained(output_dir)
    print("tokenizer saved at : " + output_dir)
    end = timer()
    print("time elapsed : {:.2f}s".format(end-start)) 

tokenizer = AutoTokenizer.from_pretrained("facebook/bart-base")

### HDFS

generate_tokenizer(hdfs_data_dir, hdfs_file, hdfs_output_dir, tokenizer, mode='hdfs')

generate_tokenizer(hdfs_data_dir, hdfs_content_file, hdfs_output_dir + "/content", tokenizer, mode='hdfs', content=True)

### BGL

generate_tokenizer(bgl_data_dir, bgl_file, bgl_output_dir, tokenizer)

generate_tokenizer(bgl_data_dir, bgl_content_file, bgl_output_dir + "/content", tokenizer, content=True)

### Thunderbird

generate_tokenizer(tbird_data_dir, tbird_file, tbird_output_dir, tokenizer)

generate_tokenizer(tbird_data_dir, tbird_content_file, tbird_output_dir + "/content", tokenizer, content=True)

## Generate Tokenizer Logkey and Content

def generate_tokenizer(log_dir, log_file, output_dir, token, mode=None, vocab_size = 50265):
    start = timer()
    df = pd.read_csv(log_dir + log_file)
    if mode == 'hdfs':
        df = df[["EventSequence", "ContentSequence"]]
        df.rename(columns={'EventSequence': 'logkey'}, inplace=True)
        df.rename(columns={'ContentSequence': 'content'}, inplace=True)
    else :
        df = df[["EventId", "EventTemplate"]]
        df.rename(columns={'EventId': 'logkey'}, inplace=True)
        df.rename(columns={'EventTemplate': 'content'}, inplace=True)
    
    dataset = Dataset.from_pandas(df)
    batch_size = 1000
    
    def batch_iterator():
        for i in range(0, len(dataset), batch_size):
            yield dataset[i : i + batch_size]["logkey"]
            yield dataset[i : i + batch_size]["content"]

    tokenizer = token.train_new_from_iterator(batch_iterator(), vocab_size)
    tokenizer.save_pretrained(output_dir + "/lc")
    print("tokenizer saved at : " + output_dir + "/lc")
    end = timer()
    print("time elapsed : {:.2f}s".format(end-start)) 

tokenizer = AutoTokenizer.from_pretrained("facebook/bart-base")

generate_tokenizer(hdfs_data_dir, hdfs_logkey_content, hdfs_output_dir, tokenizer, mode='hdfs')

generate_tokenizer(bgl_data_dir, bgl_logkey_content, bgl_output_dir, tokenizer)

generate_tokenizer(tbird_data_dir, tbird_logkey_content, tbird_output_dir, tokenizer)

## Sandbox

df = pd.read_csv("datasets/hdfs/output/hdfs_labeled.csv")

df = df.drop("BlockId", axis='columns')
df = df.drop("Label", axis='columns')

df = df[["EventSequence"]]

df.rename(columns={'EventSequence': 'text'}, inplace=True)

df.head()

dataset = Dataset.from_pandas(df)

dataset

batch_size = 1000
all_texts = [dataset[i : i + batch_size]["text"] for i in range(0, len(dataset), batch_size)]

def batch_iterator():
    for i in range(0, len(dataset), batch_size):
        yield dataset[i : i + batch_size]["text"]

tokenizer = AutoTokenizer.from_pretrained("facebook/bart-base")

tokenizer.is_fast

new_tokenizer = tokenizer.train_new_from_iterator(batch_iterator(),vocab_size = 50265)

new_tokenizer(dataset[1]["text"])

new_tokenizer.save_pretrained("tokenizer/hdfs")
