==> 04. BART Training.ipynb <==
!pip install datasets evaluate accelerate -U

import pandas as pd
import numpy as np
import torch
import zipfile
import evaluate
import json
from datasets import Dataset, load_metric
from transformers import BartTokenizerFast
from transformers import BartForSequenceClassification, Trainer, TrainingArguments, EvalPrediction, AutoConfig

train_df = pd.read_csv("/content/train_hdfs_labeled.csv")
eval_df = pd.read_csv("/content/eval_hdfs_labeled.csv")

train_df["label"].value_counts()

eval_df["label"].value_counts()

train_ds = Dataset.from_pandas(train_df)
eval_ds = Dataset.from_pandas(eval_df)

with zipfile.ZipFile("/content/hdfs.zip", 'r') as zip_ref:
  zip_ref.extractall("tokenizer")

tokenizer = tokenizer = BartTokenizerFast.from_pretrained("/content/tokenizer")

def preprocess_token(dataset):
  encoded_sequence = tokenizer(dataset["text"], truncation = True, padding = 'max_length')
  encoded_sequence['label'] = dataset["label"]
  return encoded_sequence

train_dataset = train_ds.map(preprocess_token, batched = True, batch_size = 1, remove_columns = ["text"])
eval_dataset = eval_ds.map(preprocess_token, batched = True, batch_size = 1, remove_columns = ["text"])

train_dataset.set_format("torch", columns=["input_ids", "attention_mask", "label"])
eval_dataset.set_format("torch", columns=["input_ids", "attention_mask", "label",])

eval_dataset[11]

def compute_metrics(p: EvalPrediction):
    # metric = evaluate.load("confusion_matrix")
    metric_acc = evaluate.load("accuracy")
    metric_f1 = evaluate.load("f1")
    preds = p.predictions[0] if isinstance(p.predictions, tuple) else p.predictions
    preds = np.argmax(preds, axis = 1)
    result = {}
    result["accuracy"] = metric_acc.compute(predictions = preds, references = p.label_ids)["accuracy"]
    result["f1"] = metric_f1.compute(predictions = preds, references = p.label_ids, average = 'macro')["f1"]
    # result["confusion_matrix"] = metric.compute(references=p.label_ids, predictions=preds)
    return result

id2label = {0: "Normal", 1: "Anomaly"}
label2id = {"Normal": 0, "Anomaly": 1}

training_args = TrainingArguments(
  output_dir = "models/",      # Output directory
  learning_rate=5e-5,
  num_train_epochs = 1,             # Total number of training epochs
  per_device_train_batch_size = 1,  # Batch size per device during training
  per_device_eval_batch_size = 1,   # Batch size for evaluation
  warmup_steps = 500,                # Number of warmup steps for learning rate scheduler
  weight_decay = 0.01,               # Strength of weight decay
  gradient_accumulation_steps = 4,
  gradient_checkpointing=True,
  evaluation_strategy="epoch",
  save_strategy="epoch",
  load_best_model_at_end=True,
)

model = BartForSequenceClassification.from_pretrained("facebook/bart-base", num_labels = 2, id2label=id2label, label2id=label2id, ignore_mismatched_sizes = True)

trainer = Trainer(
  model = model,                     # The instantiated model to be trained
  args = training_args,              # Training arguments, defined above
  compute_metrics = compute_metrics, # A function to compute the metrics
  train_dataset = train_dataset,     # Training dataset
  eval_dataset = eval_dataset,       # Evaluation dataset
  tokenizer = tokenizer              # The tokenizer that was used
)

trainer.train()

trainer.evaluate()

import os
import zipfile

def get_all_file_paths(directory):

    # initializing empty file paths list
    file_paths = []

    # crawling through directory and subdirectories
    for root, directories, files in os.walk(directory):
        for filename in files:
            # join the two strings in order to form the full filepath.
            filepath = os.path.join(root, filename)
            file_paths.append(filepath)

    # returning all file paths
    return file_paths

directory = '/content/models/checkpoint-1500'

    # calling function to get all file paths in the directory
file_paths = get_all_file_paths(directory)

print('Following files will be zipped:')
for file_name in file_paths:
  print(file_name)

# ZipFile = zipfile.ZipFile
# with ZipFile('/content/drive/MyDrive/Workspace/model.zip','w') as zip:
#   for file in file_paths:
#     zip.write(file)
